backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
google-genai>=1.0.0
pydantic>=2.7.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
python-pptx>=0.6.21
pywin32>=306
